backend/requirements.txt
fastapi
uvicorn
pydantic
openai
httpx
python-dotenv
pytest==8.0.2
python-multipart==0.0.9
PyPDF2==3.0.1
python-docx==1.1.0
